refactor: Move Tree.tsx into utility-components directory
