+++ packages/tools/devtools/devtools-view/src/components/FluidTreeView.tsx
@@ -4,7 +4,7 @@
  */
 import React from "react";
 import { HasContainerId, FluidObjectTreeNode } from "@fluid-experimental/devtools-core";
-import { Tree } from "./Tree";
+import { Tree } from "./utility-components";
 import { TreeDataView } from "./TreeDataView";
 import { TreeHeader } from "./TreeHeader";
 import { HasLabel } from "./CommonInterfaces";
